azure-function-databricks-monitor/requirements.txt
# Azure Functions requirements
azure-functions>=1.18.0
requests>=2.31.0
